Added permission check for bluetooth
